requirements.txt
streamlit==1.39.0
google-generativeai==0.8.3
faiss-cpu==1.12.0
pypdf==5.0.1
python-dotenv==1.0.1
fastapi==0.115.5
uvicorn==0.32.0
numpy==2.1.2
